requirements.txt
mysql-connector-python==9.1.0
prettytable==3.12.0
pyodbc==5.2.0